requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
idna==3.10
newsapi-python==0.2.7
pi==0.1.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
